Persist order line-item dimensions and price them server-side
Dimension-priced items (LFP prints etc.) previously computed area and
looked up the price tier entirely client-side; the length/breadth that
produced the number were never saved, only the resulting qty. Reopening
such an item for edit showed blank dimension fields with no way to
recover what was actually measured, and the backend blindly trusted
whatever qty/unit_price the browser posted.

order_items now has length/breadth/dimension_unit columns. The backend
recomputes area (converting inches to sqft via /144, same formula used
throughout the app) and the matching price tier itself for any item
whose product/service has requires_dimensions, ignoring the client's
qty/unit_price for those rows. Create/Edit still compute a live preview
client-side for UX, but the server is now the source of truth, and Edit
reloads the real persisted dimensions instead of starting blank.

+++ app/Models/Product.php
@@ -108,7 +108,7 @@ class Product extends Model
         return $this->calculateCost();
     }
 
-    public function getPriceForQuantity(int $quantity): ?float
+    public function getPriceForQuantity(float $quantity): ?float
     {
         $price = $this->prices()
             ->where('min_qty', '<=', $quantity)
